feat: 增加独立窗口编辑能力

+++ packages/web/src/views/desktop/main-container/pages/main-container-default/right-bar.tsx
@@ -2,10 +2,9 @@ import IconNavMenu from '@/components/icon-nav-menu';
 import showShareNoteDialog from '@/components/note-share';
 import { Note } from '@/services/note';
 import { copyText } from '@/utils';
-import { downloadFile } from '@/utils/file';
+import is from '@/utils/is';
 import { PropType, defineComponent } from 'vue';
 import './index.scss';
-import { useRouter } from 'vue-router';
 
 const DesktopMainContainerDefaultRight = defineComponent({
   name: 'DesktopMainContainerDefaultRight',
@@ -16,7 +15,6 @@ const DesktopMainContainerDefaultRight = defineComponent({
     }
   },
   setup(props) {
-    const router = useRouter();
     const menuList = [
       {
         title: '预览',
@@ -42,13 +40,15 @@ const DesktopMainContainerDefaultRight = defineComponent({
         icon: 'item-json-download',
         visible: true,
         action: (note: Note) => {
-          note && downloadFile(JSON.stringify(note) || '', `${note.title || 'XYNote'}.json`);
+          if (note) {
+            note.toJson();
+          }
         }
       },
       {
         title: '分屏',
         icon: 'item-splitscreen',
-        visible: true,
+        visible: is.app(),
         action: (note: Note) => {
           if (note) {
             handleSplitScreen(note);
@@ -60,32 +60,26 @@ const DesktopMainContainerDefaultRight = defineComponent({
         icon: 'item-delete',
         visible: true,
         action: (note: Note) => {
-          note &&
+          if (note) {
             window.$ui.confirm({
               type: 'warn',
-              width: 300,
               content: `确定删除《${note.title}》这个笔记吗？`,
               onSubmit: () => {
                 note.delete();
               }
             });
+          }
         }
       }
     ];
 
     // 新开页面
     const handleSplitScreen = (note: Note) => {
-      let routeUrl = router.resolve({
-        path: '/detail',
-        query: {
-          nid: note.nid
-        }
-      });
-      window.open(routeUrl.href, '_blank');
+      window.createWindow({ nid: note.nid });
     };
 
     return () => (
-      <div class="desktop-main-container-default-content-right">
+      <div class="desktop-main-container-default-content-right" data-tauri-drag-region>
         {menuList.map((menu) => {
           return (
             <div
